pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kibad-llm"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Your name (or your organization/company/team)" }]
requires-python = ">=3.10.0"
readme = "README.md"
classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
    "pandas>=2.3.3", # for zotero_download.py
    "pandas-stubs>=2.3.2.250926", # for zotero_download.py
    "psycopg2-binary>=2.9.10", # for db_converter.py
    "pyyaml>=6.0.3", # for db_converter.py
    "requests>=2.32.5", # for zotero_download.py
    "retry>=0.9.2", # for zotero_download.py
    "types-psycopg2>=2.9.21.20250915", # for db_converter.py
    "types-pyyaml>=6.0.12.20250915", # for db_converter.py
    "types-retry>=0.9.9.20250322", # for zotero_download.py
    "defusedxml>=0.7.1", # for synch_nextcloud_with_cluster.py
    "types-defusedxml>=0.7.0.20250822", # for synch_nextcloud_with_cluster.py
    "pymupdf4llm>=0.0.27", # pdf to markdown conversion
    "llama-index-llms-openai-like>=0.5.3", # for LLM (gpt-oss) integration
    "jsonschema>=4.25.1",  # for structured output schema validation
    "types-jsonschema>=4.25.1.20251009",
    # utils & other
    "tqdm>=4.67.1",
    "types-tqdm>=4.67.0.20250809",
    "typer>=0.19.2",
    "python-dotenv>=1.1.1",
    "loguru>=0.7.3",
    "mkdocs>=1.6.1",
    "hydra-core>=1.3.2",
    "hydra-colorlog>=1.2.0",
    "datasets<4.0.0",
    "pdfplumber>=0.11.7",
]

[dependency-groups]
cicd = [
    "pytest>=6.2.5",
    "pytest-xdist>=3.8.0,<4",
    "pytest-cov>=6.2.1,<7",
    "typeguard>=2.13.3",
    "pre-commit>=2.16.0",
    "mypy>=1.3.0",
]

[tool.hatch.build.targets.sdist]
include = ["src/kibad_llm"]

[tool.hatch.build.targets.wheel]
include = ["src/kibad_llm"]

[tool.hatch.build.targets.wheel.sources]
"src/kibad_llm" = "kibad_llm"

[tool.pytest.ini_options]
typeguard-packages = "kibad_llm"
minversion = "6.0"
testpaths = ["tests"]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "typeguard_ignore: ignore typeguard in this test"
]

[tool.coverage.run]
branch = true
source = ["kibad_llm"]   # measure only your package (src layout is fine)

[tool.coverage.report]
show_missing = true

[tool.black]
line-length = 99
# limit Black’s directory walk to real Python files only (.py / .pyi);
# avoids touching other extensions and speeds discovery on large repos
include = '\.pyi?$'
# don’t even scan common “noise” dirs (.git, virtualenv) when walking the tree;
# keeps runs fast and prevents accidental formatting of vendored/ephemeral files.
# triple-quoted here on purpose: Black treats multiline patterns as “verbose”
# regex, making it easy to extend later.
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.isort]
profile = "black"
known_first_party = ["kibad_llm"]
force_sort_within_sections = true

[tool.flake8]
ignore = ["E203", "E266", "E402", "E501", "F401", "F841", "C901", "W503", "E731"]
max-line-length = 99
exclude = [".git", "notebooks", "models", "data", "logs"]

[tool.codespell]
skip = "logs/**,data/**,configs/template/**,tests/fixtures/**.json,src/kibad_llm/schema/**,tests/unit/schema/**,*.ipynb"
ignore-words-list = "hist"

# Bandit (read by pre-commit bandit with -c pyproject.toml)
[tool.bandit]
skips = ["B101", "B107", "B403"]
# optionally exclude tests to reduce noise:
# exclude_dirs = ["tests"]

# mdformat (via mdformat-pyproject); tweak if you want numbering/wrapping
[tool.mdformat]
wrap = "keep"          # {"keep","no",INTEGER}
number = false         # consecutive numbering for ordered lists
end_of_line = "lf"
validate = true

# (optional) nbQA-specific knobs if you need them later:
# [tool.nbqa.config]
# black = "pyproject.toml"
# isort = "pyproject.toml"
# flake8 = "pyproject.toml"

[tool.mypy]
python_version = "3.10"
files = ["src", "tests"]

[[tool.mypy.overrides]]
module = ["pymupdf4llm", "datasets"]
ignore_missing_imports = true
